delete create

+++ src/main/java/bip/online/biplio2023/repo/AuthorRepo.java
@@ -3,6 +3,6 @@ package bip.online.biplio2023.repo;
 import bip.online.biplio2023.entity.Author;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuthorRepo extends JpaRepository<Author, Integer> {
+public interface AuthorRepo extends JpaRepository<Author, Long> {
 
 }
