+++ src/main/java/bip/online/biplio2023/repo/CityRepo.java
@@ -3,6 +3,6 @@ package bip.online.biplio2023.repo;
 import bip.online.biplio2023.entity.City;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CityRepo extends JpaRepository<City, Integer> {
+public interface CityRepo extends JpaRepository<City, Long> {
 
 }
