+++ src/main/java/bip/online/biplio2023/repo/PublisherRepo.java
@@ -3,5 +3,5 @@ package bip.online.biplio2023.repo;
 import bip.online.biplio2023.entity.Publisher;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PublisherRepo extends JpaRepository<Publisher, Integer> {
+public interface PublisherRepo extends JpaRepository<Publisher, Long> {
 }
